deploy/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.6.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.5
hexbytes==0.3.1
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.10.0
protobuf==5.27.0
pycryptodome==3.20.0
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rlp==4.0.1
rpds-py==0.18.1
sniffio==1.3.1
toolz==0.12.1
typing_extensions==4.12.0
urllib3==2.2.1
web3==6.19.0
websockets==12.0
yarl==1.9.4
